# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SynthScribe API — turn a hummed melody into a fully produced song.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * The musical mood/genre direction for the backing track.
 */
export type Vibe = typeof Vibe[keyof typeof Vibe];


export const Vibe = {
  lofi: 'lofi',
  cinematic: 'cinematic',
  pop: 'pop',
  rnb: 'rnb',
  electronic: 'electronic',
  acoustic: 'acoustic',
  ambient: 'ambient',
  serenity: 'serenity',
  soul: 'soul',
  jazz: 'jazz',
  folk: 'folk',
  afrobeat: 'afrobeat',
  synthwave: 'synthwave',
} as const;

/**
 * Which engine produces the backing track. "musicgen" (default) is the Modal MusicGen-melody GPU worker, conditioned on the hum so the band follows the real tune. "elevenlabs" uses the premium ElevenLabs Music model. "arranger" and "gpu" are legacy values kept only so older projects still deserialize; they are not offered in the UI.
 */
export type Engine = typeof Engine[keyof typeof Engine];


export const Engine = {
  musicgen: 'musicgen',
  elevenlabs: 'elevenlabs',
  arranger: 'arranger',
  gpu: 'gpu',
} as const;

/**
 * Engine a user may choose when creating a project. Only the two currently-offered engines are valid; legacy values ("arranger", "gpu") are not accepted for new projects.
 */
export type EngineChoice = typeof EngineChoice[keyof typeof EngineChoice];


export const EngineChoice = {
  musicgen: 'musicgen',
  elevenlabs: 'elevenlabs',
} as const;

/**
 * Current stage of the hum-to-song pipeline.
 */
export type PipelineStage = typeof PipelineStage[keyof typeof PipelineStage];


export const PipelineStage = {
  draft: 'draft',
  transcribing: 'transcribing',
  generating_backing: 'generating_backing',
  singing: 'singing',
  mixing: 'mixing',
  complete: 'complete',
  error: 'error',
} as const;

export interface CreateProjectRequest {
  title?: string;
  vibe: Vibe;
  engine?: EngineChoice;
}

/**
 * Public URLs for each produced audio stage, when available.
 */
export interface AudioUrls {
  /** @nullable */
  hum: string | null;
  /** @nullable */
  backing: string | null;
  /** @nullable */
  vocals: string | null;
  /** @nullable */
  final: string | null;
}

export interface Project {
  id: string;
  title: string;
  vibe: Vibe;
  engine: Engine;
  stage: PipelineStage;
  /** Coarse progress percentage (0-100). */
  progress: number;
  /** @nullable */
  key?: string | null;
  /** @nullable */
  tempo?: number | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  error?: string | null;
  audio: AudioUrls;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectStatus {
  id: string;
  stage: PipelineStage;
  progress: number;
  /**
     * Human-friendly status line, e.g. "Warming up the music model".
     * @nullable
     */
  message?: string | null;
  /** @nullable */
  error?: string | null;
  audio: AudioUrls;
}

export interface ErrorResponse {
  error: string;
}

export type UploadHumBody = {
  file: Blob;
};