# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SynthScribe API — turn a hummed melody into a fully produced song.
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Song project operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List recent projects
      responses:
        "200":
          description: A list of projects, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new song project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by id
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: The project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/hum:
    post:
      operationId: uploadHum
      tags: [projects]
      summary: Upload the hummed recording for a project
      description: Accepts a 44.1kHz WAV recording of the user's hum.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Hum stored, project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/generate:
    post:
      operationId: startGeneration
      tags: [projects]
      summary: Start the hum-to-song pipeline
      description: Kicks off transcription, backing track, vocals, and mixing in the background.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "202":
          description: Generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: No hum uploaded yet, or already generating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/status:
    get:
      operationId: getProjectStatus
      tags: [projects]
      summary: Poll the pipeline status for a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Current pipeline status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Vibe:
      type: string
      description: The musical mood/genre direction for the backing track.
      enum:
        - lofi
        - cinematic
        - pop
        - rnb
        - electronic
        - acoustic
        - ambient
        - serenity
        - soul
        - jazz
        - folk
        - afrobeat
        - synthwave
    Engine:
      type: string
      description: Which engine produces the backing track. "musicgen" (default) is the Modal MusicGen-melody GPU worker, conditioned on the hum so the band follows the real tune. "elevenlabs" uses the premium ElevenLabs Music model. "arranger" and "gpu" are legacy values kept only so older projects still deserialize; they are not offered in the UI.
      enum:
        - musicgen
        - elevenlabs
        - arranger
        - gpu
    EngineChoice:
      type: string
      description: Engine a user may choose when creating a project. Only the two currently-offered engines are valid; legacy values ("arranger", "gpu") are not accepted for new projects.
      enum:
        - musicgen
        - elevenlabs
    PipelineStage:
      type: string
      description: Current stage of the hum-to-song pipeline.
      enum:
        - draft
        - transcribing
        - generating_backing
        - singing
        - mixing
        - complete
        - error
    CreateProjectRequest:
      type: object
      properties:
        title:
          type: string
        vibe:
          $ref: "#/components/schemas/Vibe"
        engine:
          $ref: "#/components/schemas/EngineChoice"
      required:
        - vibe
    AudioUrls:
      type: object
      description: Public URLs for each produced audio stage, when available.
      properties:
        hum:
          type: [string, "null"]
        backing:
          type: [string, "null"]
        vocals:
          type: [string, "null"]
        final:
          type: [string, "null"]
      required:
        - hum
        - backing
        - vocals
        - final
    Project:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        vibe:
          $ref: "#/components/schemas/Vibe"
        engine:
          $ref: "#/components/schemas/Engine"
        stage:
          $ref: "#/components/schemas/PipelineStage"
        progress:
          type: integer
          description: Coarse progress percentage (0-100).
        key:
          type: [string, "null"]
        tempo:
          type: [number, "null"]
        durationSeconds:
          type: [number, "null"]
        error:
          type: [string, "null"]
        audio:
          $ref: "#/components/schemas/AudioUrls"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - vibe
        - engine
        - stage
        - progress
        - audio
        - createdAt
        - updatedAt
    ProjectStatus:
      type: object
      properties:
        id:
          type: string
        stage:
          $ref: "#/components/schemas/PipelineStage"
        progress:
          type: integer
        message:
          type: [string, "null"]
          description: Human-friendly status line, e.g. "Warming up the music model".
        error:
          type: [string, "null"]
        audio:
          $ref: "#/components/schemas/AudioUrls"
      required:
        - id
        - stage
        - progress
        - audio
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error