# lib/api-zod/src/generated/types/createProjectRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SynthScribe API — turn a hummed melody into a fully produced song.
 * OpenAPI spec version: 0.1.0
 */
import type { EngineChoice } from './engineChoice';
import type { Vibe } from './vibe';

export interface CreateProjectRequest {
  title?: string;
  vibe: Vibe;
  engine?: EngineChoice;
}