# lib/api-zod/src/generated/types/engine.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SynthScribe API — turn a hummed melody into a fully produced song.
 * OpenAPI spec version: 0.1.0
 */

/**
 * Which engine produces the backing track. "musicgen" (default) is the Modal MusicGen-melody GPU worker, conditioned on the hum so the band follows the real tune. "elevenlabs" uses the premium ElevenLabs Music model. "arranger" and "gpu" are legacy values kept only so older projects still deserialize; they are not offered in the UI.
 */
export type Engine = typeof Engine[keyof typeof Engine];


export const Engine = {
  musicgen: 'musicgen',
  elevenlabs: 'elevenlabs',
  arranger: 'arranger',
  gpu: 'gpu',
} as const;