# lib/api-zod/src/generated/types/engineChoice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SynthScribe API — turn a hummed melody into a fully produced song.
 * OpenAPI spec version: 0.1.0
 */

/**
 * Engine a user may choose when creating a project. Only the two currently-offered engines are valid; legacy values ("arranger", "gpu") are not accepted for new projects.
 */
export type EngineChoice = typeof EngineChoice[keyof typeof EngineChoice];


export const EngineChoice = {
  musicgen: 'musicgen',
  elevenlabs: 'elevenlabs',
} as const;